+++ OmeSliCC/conversion.py
@@ -1,7 +1,5 @@
 # https://pyquestions.com/how-to-save-a-very-large-numpy-array-as-an-image-loading-as-little-as-possible-into-memory
-# * TODO: fix Zarr support, extend to Ome.Zarr
-# * TODO: Add JPEGXR support for Zarr
-# TODO: use ome-zarr python package for reading/writing if supporting image compressors (Jpeg2k/xr)
+
 
 import os
 import numpy as np
@@ -39,18 +37,19 @@ def create_source(source_ref: str, params: dict, omero: Omero = None) -> OmeSour
 
 
 def get_image_info(source: OmeSource) -> str:
-    xyzct = source.get_size_xyzct()
+    sizes_xyzct = source.sizes_xyzct
     pixel_nbytes = source.get_pixel_nbytes()
     pixel_type = source.get_pixel_type()
     channels = source.get_channels()
-    image_info = os.path.basename(source.source_reference) + '\n'
-    image_info += get_image_size_info(xyzct, pixel_nbytes, pixel_type, channels)
+    image_info = os.path.basename(source.source_reference)
+    image_info += ' ' + get_image_size_info(sizes_xyzct, pixel_nbytes, pixel_type, channels)
     sizes = source.get_physical_size()
     if len(sizes) > 0:
-        image_info += '\nPhysical size:'
+        image_info += ' Physical size:'
         infos = []
         for size in sizes:
-            infos.append(f' {size[0]:.3f} {size[1]}')
+            if size[0] > 0:
+                infos.append(f' {size[0]:.2f} {size[1]}')
         image_info += ' x'.join(infos)
     return image_info
 
@@ -130,7 +129,7 @@ def combine_images(sources: list[OmeSource], params: dict):
                 channel['@Name'] = input_channels[c]
         channels.append(channel)
     filetitle = get_filetitle(source_ref).rstrip('.ome')
-    output_filename = os.path.join(output_folder, filetitle + '.' + output_format)
+    output_filename = str(os.path.join(output_folder, filetitle + '.' + output_format))
     if 'zar' in output_format:
         new_source = OmeZarrSource(source_ref, source0.get_pixel_size())
         new_source.channels = channels
@@ -161,12 +160,12 @@ def get_source_image_chunked(source: OmeSource, chunk_size=(10240, 10240)):
     image = source.clone_empty()
     for indices, chunk in source.produce_chunks(chunk_size):
         s = indices
-        d = chunk.shape
-        image[s[0]:s[0] + d[0],
-              s[1]:s[1] + d[1],
-              s[2]:s[2] + d[2],
-              s[3]:s[3] + d[3],
-              s[4]:s[4] + d[4]] = chunk
+        e = np.array(s) + chunk.shape
+        image[s[0]:e[0],
+              s[1]:e[1],
+              s[2]:e[2],
+              s[3]:e[3],
+              s[4]:e[4]] = chunk
     return image
 
 
@@ -196,17 +195,19 @@ def save_image_as_zarr(source: OmeSource, data: np.ndarray, output_filename: str
 
 
 def save_image_as_tiff(source: OmeSource, image: np.ndarray, output_filename: str, output_params: dict, ome: bool = False):
-    dimension_order = source.get_dimension_order()
     tile_size = output_params.get('tile_size')
     compression = output_params.get('compression')
     combine_rgb = output_params.get('combine_rgb', True)
+
+    # move color channels to end:
+    image, dimension_order = source.get_yxc_image(image)
 
     npyramid_add = output_params.get('npyramid_add', 0)
     pyramid_downsample = output_params.get('pyramid_downsample')
     if npyramid_add > 0:
         pyramid_sizes_add = calc_pyramid(source.get_size_xyzct(), npyramid_add, pyramid_downsample)
     else:
-        pyramid_sizes_add = None
+        pyramid_sizes_add = []
 
     if ome:
         metadata = None
@@ -227,7 +228,7 @@ def save_image_as_tiff(source: OmeSource, image: np.ndarray, output_filename: st
 def save_tiff(filename: str, image: np.ndarray, metadata: dict = None, xml_metadata: str = None,
               dimension_order: str = 'yxc',
               resolution: tuple = None, resolution_unit: str = None, tile_size: tuple = None, compression: [] = None,
-              combine_rgb=True, pyramid_sizes_add: list = None):
+              combine_rgb=True, pyramid_sizes_add: list = []):
     x_index = dimension_order.index('x')
     y_index = dimension_order.index('y')
     size = image.shape[x_index], image.shape[y_index]
@@ -252,23 +253,23 @@ def save_tiff(filename: str, image: np.ndarray, metadata: dict = None, xml_metad
         # tifffile only supports x/y pyramid resolution
         resolution = tuple(resolution[0:2])
 
-    # estimate size (w/o compression)
-    estimated_size = image.size * image.itemsize
-    base_size = np.divide(estimated_size, np.prod(size))
+    # maximum size (w/o compression)
+    max_size = image.size * image.itemsize
+    base_size = np.divide(max_size, np.prod(size))
     for new_size in pyramid_sizes_add:
-        estimated_size += np.prod(new_size) * base_size
-    bigtiff = (estimated_size > 2 ** 32)
+        max_size += np.prod(new_size) * base_size
+    bigtiff = (max_size > 2 ** 32)
 
     #scaler = Scaler(downscale=..., max_layer=len(pyramid_sizes_add))   # use ome-zarr-py dask scaling
 
     # set ome=False to provide custom OME xml in description
     xml_metadata_bytes = xml_metadata.encode() if xml_metadata is not None else None
     with TiffWriter(filename, ome=False, bigtiff=bigtiff) as writer:
-        writer.write(image.squeeze(), photometric=photometric, subifds=len(pyramid_sizes_add),
+        writer.write(image, photometric=photometric, subifds=len(pyramid_sizes_add),
                      resolution=resolution, resolutionunit=resolution_unit, tile=tile_size, compression=compression,
                      metadata=metadata, description=xml_metadata_bytes)
         for new_size in pyramid_sizes_add:
             image = image_resize(image, new_size, dimension_order=dimension_order)
             #image = scaler.resize_image(image)    # significantly slower
-            writer.write(image.squeeze(), photometric=photometric, subfiletype=1,
+            writer.write(image, photometric=photometric, subfiletype=1,
                          resolution=resolution, resolutionunit=resolution_unit, tile=tile_size, compression=compression)